requirements.txt
telebot
opencv-python-headless
Pillow
pywin32
psutil
plyer
win10toast
